refactor: divide pacquet-fs into multiple files
