+++ crates/fs/src/ensure_file.rs
@@ -1,0 +1,66 @@
+use derive_more::{Display, Error};
+use miette::Diagnostic;
+use std::{
+    fs::{self, OpenOptions},
+    io::{self, Write},
+    path::{Path, PathBuf},
+};
+
+/// Error type of [`ensure_file`].
+#[derive(Debug, Display, Error, Diagnostic)]
+pub enum EnsureFileError {
+    #[display("Failed to create the parent directory at {parent_dir:?}: {error}")]
+    CreateDir {
+        parent_dir: PathBuf,
+        #[error(source)]
+        error: io::Error,
+    },
+    #[display("Failed to create file at {file_path:?}: {error}")]
+    CreateFile {
+        file_path: PathBuf,
+        #[error(source)]
+        error: io::Error,
+    },
+    #[display("Failed to write to file at {file_path:?}: {error}")]
+    WriteFile {
+        file_path: PathBuf,
+        #[error(source)]
+        error: io::Error,
+    },
+}
+
+/// Write `content` to `file_path` unless it already exists.
+///
+/// Ancestor directories will be created if they don't already exist.
+pub fn ensure_file(
+    file_path: &Path,
+    content: &[u8],
+    #[cfg_attr(windows, allow(unused))] mode: Option<u32>,
+) -> Result<(), EnsureFileError> {
+    if file_path.exists() {
+        return Ok(());
+    }
+
+    let parent_dir = file_path.parent().unwrap();
+    fs::create_dir_all(parent_dir).map_err(|error| EnsureFileError::CreateDir {
+        parent_dir: parent_dir.to_path_buf(),
+        error,
+    })?;
+
+    let mut options = OpenOptions::new();
+    options.write(true).create(true);
+
+    #[cfg(unix)]
+    {
+        use std::os::unix::fs::OpenOptionsExt;
+        if let Some(mode) = mode {
+            options.mode(mode);
+        }
+    }
+
+    options
+        .open(file_path)
+        .map_err(|error| EnsureFileError::CreateFile { file_path: file_path.to_path_buf(), error })?
+        .write_all(content)
+        .map_err(|error| EnsureFileError::WriteFile { file_path: file_path.to_path_buf(), error })
+}
